+++ src/CRSFforArduino/src/CRSFforArduino.hpp
@@ -49,7 +49,7 @@ namespace sketchLayer
         uint16_t rcToUs(uint16_t rc);
         uint16_t readRcChannel(uint8_t channel, bool raw = false);
 
-        void telemetryWriteAttitude(float roll, float pitch, float yaw);
+        void telemetryWriteAttitude(int16_t roll, int16_t pitch, int16_t yaw);
         void telemetryWriteBattery(float voltage, float current, uint32_t fuel, uint8_t percent);
         void telemetryWriteGPS(float latitude, float longitude, float altitude, float speed, float groundCourse, uint8_t satellites);
